MYR-409 Read the nav stamp and the dispatch instant in both projections
Both content-state projections — the ticker's per-pass list and the
lifecycle fan-out's single-ride read — now source NavUpdatedAt from
nav_reading_at instead of the car's row stamp, and both carry
go_ride_requests.dispatched_at as PickupDispatchedAt.

The join is LEFT and deliberately NOT COALESCEd onto the row stamp. A
missing row means no navigation frame has ever been seen, and NULL is the
honest way to say so: every gate reads a nil stamp as not fresh. Falling
back to lastUpdated would restore the exact lie the column exists to end,
for exactly the population it describes.

The single-ride read moves to live_activity_context.go — the shared join and
predicate keep the two queries from drifting, and the split keeps both files
under the 300-line cap.
